pyproject.toml
[project]
name = "plantuml-compose"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "jealouscloud", email = "github@noaha.org" }
]
requires-python = ">=3.13"
dependencies = [
    "pyyaml>=6.0",
]

[project.scripts]
plantuml-compose = "plantuml_compose:main"

[build-system]
requires = ["uv_build>=0.9.13,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
